# lib/api-zod/src/generated/types/brandConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Graphic Creator API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandConfig {
  /** Hex color */
  primaryColor: string;
  /** Hex color */
  secondaryColor: string;
  /** Hex color */
  accentColor: string;
  /** Logo image URL or base64 */
  logoUrl?: string;
  organizationName: string;
}