# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Graphic Creator API
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: number;
  name: string;
  category: string;
  /** Fabric.js canvas JSON */
  canvasJson: string;
  /** Base64 thumbnail */
  thumbnail?: string;
  createdAt: Date;
}